+++ src/media_tools/mpeg2_ps.c
@@ -157,7 +157,7 @@ static Bool file_read_bytes(FILE *fd,
 			     u8 *buffer, 
 			     u32 len)
 {
-  u32 readval = fread(buffer, 1, len, fd);
+  u32 readval = (u32) fread(buffer, 1, len, fd);
   return readval == len;
 }
